gateway_announcement: we can disable the announcement now

+++ roles/gateway_announcement/templates/gateway_announcement.sh
@@ -6,6 +6,11 @@ GW_BANDWIDTH=/sys/class/net/bat0/mesh/gw_bandwidth
 # ensure that we announce the highest bandwidth
 echo "96MBit/96MBit" > $GW_BANDWIDTH
 
+if [ "$1" = '--force-off' ]; then
+	echo off > $GW_MODE || (echo "batman gw mode failed: off"; exit 1)
+	exit 0
+fi
+
 if ping -q -I internetz-me 8.8.8.8 -c 4 -W 5 >/dev/null; then
 	echo server > $GW_MODE || logger "batman gw mode failed: server"
 	systemctl start isc-dhcp-server
